# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "RealmForge Control API",
    "description": "Governance kernel REST API — session, command, audit, snapshot, rollback, boards, and more.",
    "license": {
      "name": "UNLICENSED",
      "identifier": "UNLICENSED"
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/v1/bundles/{id}/deploy": {
      "post": {
        "tags": [
          "bundles"
        ],
        "summary": "Mark a verified bundle as deployed.",
        "operationId": "deploy_bundle",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Bundle ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Bundle marked deployed"
          },
          "404": {
            "description": "Bundle not found"
          },
          "500": {
            "description": "Internal error"
          }
        }
      }
    },
...
    "/v1/runtime": {
      "get": {
        "tags": [
          "runtime"
        ],
        "summary": "List all active runtime instances.",
        "operationId": "list_runtimes",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "format": "int64"
            },
            "example": 20
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "format": "int64"
            },
            "example": 0
          }
        ],
        "responses": {
          "200": {
            "description": "Runtime instance list"
          },
          "500": {
            "description": "Internal error"
          }
        }
      }
    },
    "/v1/runtime/deploy": {
      "post": {
        "tags": [
          "runtime"
        ],
        "summary": "Deploy a bundle as a new runtime instance.",
        "operationId": "deploy_runtime",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeployBundleRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Runtime deployed"
          },
          "400": {
            "description": "Invalid request"
          },
          "500": {
            "description": "Internal error"
          }
        }
      }
    },
    "/v1/runtime/{id}": {
      "get": {
        "tags": [
          "runtime"
        ],
        "summary": "Get the current status of a runtime instance.",
        "operationId": "get_runtime_status",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Runtime ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Runtime instance status"
          },
          "404": {
            "description": "Runtime not found"
          },
          "500": {
            "description": "Internal error"
          }
        }
      }
    },
    "/v1/runtime/{id}/stop": {
      "post": {
        "tags": [
          "runtime"
        ],
        "summary": "Stop a running runtime instance.",
        "operationId": "stop_runtime",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Runtime ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Runtime stopped"
          },
          "404": {
            "description": "Runtime not found"
          },
          "500": {
            "description": "Internal error"
          }
        }
      }
    },
    "/v1/session": {
      "post": {
        "tags": [
          "sessions"
        ],
        "summary": "Issue a new session for an actor within a tenant and project.",
        "operationId": "issue_session",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IssueSessionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Session issued",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/v1/session/{id}": {
      "get": {
        "tags": [
          "sessions"
        ],
        "summary": "Get session details by ID.",
        "operationId": "get_session",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Session ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Session found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            }
          },
          "404": {
            "description": "Session not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "sessions"
        ],
        "summary": "Revoke a session, preventing further use.",
        "operationId": "revoke_session",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Session ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Session revoked",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RevokeSessionResponse"
                }
              }
            }
          },
          "404": {
            "description": "Session not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/v1/session/{id}/activate": {
      "post": {
        "tags": [
          "sessions"
        ],
        "summary": "Activate an issued session, transitioning it from issued to active.",
        "operationId": "activate_session",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Session ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Session activated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid session state transition",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Session not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/v1/session/{id}/renew": {
      "post": {
        "tags": [
          "sessions"
        ],
        "summary": "Extend a session's TTL.",
        "operationId": "renew_session",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Session ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RenewSessionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Session renewed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            }
          },
          "404": {
            "description": "Session not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DeployBundleRequest": {
        "type": "object",
        "required": [
          "runtime_id",
          "bundle_id",
          "metadata"
        ],
        "properties": {
          "active_sessions": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "bundle_id": {
            "type": "string",
            "example": "bundle_01"
          },
          "metadata": {
            "type": "object"
          },
          "runtime_id": {
            "type": "string",
            "example": "runtime_01"
          }
        }
      },
...
      "IssueSessionRequest": {
        "type": "object",
        "required": [
          "tenant_id",
          "project_id",
          "actor_id",
          "ttl_seconds"
        ],
        "properties": {
          "actor_id": {
            "type": "string",
            "example": "alice"
          },
          "project_id": {
            "type": "string",
            "example": "proj_01"
          },
          "tenant_id": {
            "type": "string",
            "example": "actor_01"
          },
          "ttl_seconds": {
            "type": "integer",
            "format": "int64",
            "example": 3600
          }
        }
      },
...
      "ProblemDetails": {
        "type": "object",
        "description": "RFC 7807 Problem Details response, extended with a machine-readable error code\nand trace ID per the RealmForge API error standard.",
        "required": [
          "code",
          "title",
          "status",
          "detail"
        ],
        "properties": {
          "code": {
            "type": "string",
            "description": "Machine-readable SCREAMING_SNAKE_CASE error code. Stable across versions."
          },
          "detail": {
            "type": "string"
          },
          "instance": {
            "type": [
              "string",
              "null"
            ]
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "title": {
            "type": "string"
          },
          "trace_id": {
            "type": [
              "string",
              "null"
            ],
            "description": "Correlation ID matching the X-Request-Id response header."
          }
        }
      },
...
      "RenewSessionRequest": {
        "type": "object",
        "required": [
          "ttl_seconds"
        ],
        "properties": {
          "ttl_seconds": {
            "type": "integer",
            "format": "int64",
            "example": 3600
          }
        }
      },
      "RevokeSessionResponse": {
        "type": "object",
        "required": [
          "status",
          "session_id"
        ],
        "properties": {
          "session_id": {
            "type": "string"
          },
          "status": {
            "type": "string"
          }
        }
      },
...
      "SessionResponse": {
        "type": "object",
        "required": [
          "session"
        ],
        "properties": {
          "session": {
            "type": "object"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "session_token"
      }
    }
  },
...
}